server/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.2.0
dateparser==1.1.6
Deprecated==1.2.13
Django==4.1.7
django-admin-interface==0.20.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-colorfield==0.7.3
django-cors-headers==3.13.0
django-enum-choices==2.1.4
django-environ==0.9.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-redis==5.2.0
django-rest-auth==0.9.5
django-storages==1.13.1
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
exceptiongroup==1.1.0
expression==4.2.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
flower==1.2.0
goose3==3.1.12
gunicorn==20.1.0
humanize==4.6.0
idna==3.4
jieba3k==0.35.1
joblib==1.2.0
kombu==5.2.4
langdetect==1.0.9
lxml==4.9.1
newspaper3k==0.2.8
nltk==3.8.1
packaging==23.0
Pillow==8.3.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.1
pyahocorasick==2.0.0
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.2.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tinysegmenter==0.3
tldextract==3.4.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.1
